+++ RELEASENOTES.md
@@ -60,3 +60,5 @@
 - Fix A_Stop.
 - Fix RIP flag doing more damage when speed forces object into multiple smaller moves in one tick.
 - Fix A_CheckAmmo to not be inclusive.
+- Fix default args for A_MonsterBulletAttack, A_MonsterMeleeAttack, A_WeaponBulletAttack, and A_WeaponMeleeAttack.
+- Fix dehacked parser incorrectly clearing bits when encountering Bits2 and Bits3.